pyproject.toml
[project]
name = "volatilespace"
version = "0.5.2"
description = "2D physics-based universe and spaceflight simulator. "
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "pygame-ce>=2.5.5",
    "pywin32>=310 ; sys_platform == 'win32'",
]

[dependency-groups]
build = [
    "pyinstaller>=6.14.1",
]
